--- a/.github/.cspell/project-dictionary.txt
+++ b/.github/.cspell/project-dictionary.txt
@@ -10,6 +10,7 @@ emin
 endfacet
 endloop
 endsolid
+mmap
 gltf
 idents
 IDREF