--- a/.github/.cspell/rust-dependencies.txt
+++ b/.github/.cspell/rust-dependencies.txt
@@ -3,7 +3,6 @@
 
 assimp
 fastrand
-memchr
 quickcheck
 roxmltree
 walkdir