--- a/.github/.cspell/rust-dependencies.txt
+++ b/.github/.cspell/rust-dependencies.txt
@@ -3,6 +3,8 @@
 
 assimp
 fastrand
+lexopt
+memmap
 quickcheck
 roxmltree
 walkdir