… example/examples/kiss3d.rs
… tests/assimp.rs